moved things around a bit
