+++ app/src/main/java/pl/lambada/songsync/ui/screens/home/components/HomeTopAppBarDropDown.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.datastore.preferences.core.stringPreferencesKey
 import pl.lambada.songsync.R
-import pl.lambada.songsync.ui.screens.about.Providers
+import pl.lambada.songsync.util.Providers
 import pl.lambada.songsync.util.dataStore
 import pl.lambada.songsync.util.set
